vista de mis productos con las implementaciones de videojuego y pelicula

+++ src/main/java/org/springframework/samples/petclinic/web/PeliculaController.java
@@ -1,6 +1,5 @@
 package org.springframework.samples.petclinic.web;
 
-
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
@@ -10,11 +9,18 @@ import java.util.Map;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.samples.petclinic.model.Merchandasing;
 import org.springframework.samples.petclinic.model.Owner;
 import org.springframework.samples.petclinic.model.Pelicula;
+import org.springframework.samples.petclinic.model.Vendedor;
+import org.springframework.samples.petclinic.model.Videojuego;
 import org.springframework.samples.petclinic.service.PedidoService;
 import org.springframework.samples.petclinic.service.PeliculaService;
 import org.springframework.samples.petclinic.service.ProductoService;
+import org.springframework.samples.petclinic.service.VendedorService;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.ModelMap;
@@ -34,71 +40,75 @@ import net.bytebuddy.implementation.bind.MethodDelegationBinder.BindingResolver;
  */
 @Controller
 public class PeliculaController {
-	
+
 	@Autowired
 	private final PeliculaService peliculaService;
-	
+
 	@Autowired
 	private final ProductoService productoService;
-	
+
 	@Autowired
 	private final PedidoService pedidoService;
-	
+
+	@Autowired
+	private final VendedorService vendedorService;
+
 	@InitBinder
 	public void setAllowedFields(WebDataBinder dataBinder) {
 		dataBinder.setDisallowedFields("id");
 	}
 
-	
 	@Autowired
-	public PeliculaController(final PeliculaService peliculaService,final ProductoService productoService, final PedidoService pedidoService) {
+	public PeliculaController(final PeliculaService peliculaService, final ProductoService productoService,
+			final PedidoService pedidoService, final VendedorService vendedorService) {
 		this.peliculaService = peliculaService;
 		this.productoService = productoService;
 		this.pedidoService = pedidoService;
-		
+		this.vendedorService = vendedorService;
+
 	}
+
 	
-	//Muestra todas las peliculas
 	@GetMapping(value = "/peliculas")
 	public String showPeliculasList(Map<String, Object> model) {
 		List<Pelicula> peliculas = this.peliculaService.findPeliculas();
 		List<Integer> idProhibidos = this.pedidoService.listaIdPeliculasCompradas();
 		List<Pelicula> peliculasPermitidas = new ArrayList<>();
-		for(Pelicula pel: peliculas) {
-			if(!idProhibidos.contains(pel.getId())) {
+		for (Pelicula pel : peliculas) {
+			if (!idProhibidos.contains(pel.getId())) {
 				peliculasPermitidas.add(pel);
 			}
 		}
 		model.put("peliculas", peliculasPermitidas);
 		return "/peliculas/PeliculasList";
 	}
-	
-	
+
 	@GetMapping(value = "/peliculas/edit/{peliculaId}")
-	public String initUpdateForm(@PathVariable("peliculaId") int peliculaId, ModelMap model){
-		Pelicula p= this.peliculaService.findPeliculaById(peliculaId);
+	public String initUpdateForm(@PathVariable("peliculaId") int peliculaId, ModelMap model) {
+		Pelicula p = this.peliculaService.findPeliculaById(peliculaId);
 		model.put("pelicula", p);
-		String view= "/peliculas/formCreatePeliculas";
+		String view = "/peliculas/formCreatePeliculas";
 		return view;
 	}
-	
-	@PostMapping(value ="/peliculas/edit/{peliculaId}")
+
+	@PostMapping(value = "/peliculas/edit/{peliculaId}")
 	public String processUpdatePeliculaForm(@Valid final Pelicula p, BindingResult result,
-			@PathVariable("peliculaId") int peliculaId) {
-		
+			@PathVariable("peliculaId") int peliculaId,ModelMap model) {
+
 		String view = "/peliculas/formCreatePeliculas";
-		if(result.hasErrors()) {
-			return view;
-		}else {
+
+		if (result.hasErrors()) {
+			model.addAttribute("message", "ERROR!");
+		} else {
 			p.setId(peliculaId);
 			this.peliculaService.savePelicula(p);
-			view ="redirect:/peliculas/" + p.getId() ;
-			return view;
+			view = "redirect:/peliculas/" + p.getId();
+			
 		}
-		
+		return view;
 	}
+
 	
-	//Muestra en la vista peliculaDetails los detalles de una pelicula pasando su id
 	@GetMapping("/peliculas/{peliculaId}")
 	public String showPelicula(@PathVariable("peliculaId") int peliculaId, Map<String, Object> model) {
 		Pelicula pelicula = this.peliculaService.findPeliculaById(peliculaId);
@@ -106,39 +116,89 @@ public class PeliculaController {
 		return "/peliculas/peliculaDetails";
 	}
 
-	//Muestra la vista para crear una pelicula
-	@GetMapping(value= "/peliculas/new")
+
+	@GetMapping(value = "/peliculas/new")
 	public String createPelicula(final ModelMap modelmap) {
 		String view = "/peliculas/formCreatePeliculas";
-		modelmap.addAttribute("pelicula", new Pelicula());
+		modelmap.addAttribute("pelicula", new Pelicula()); 
 		return view;
 	}
+
 	
-	//Se traen los datos de la pelicula nueva del form y se procede al post 
-	@PostMapping(value= "/peliculas/new")
-	public String savePelicula(@Valid Pelicula pel, BindingResult result) {
-		if(result.hasErrors()) {
+	@PostMapping(value = "/peliculas/new")
+	public String savePelicula(@Valid Pelicula pel, BindingResult result, ModelMap model) {
+
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		UserDetails userDetail = (UserDetails) auth.getPrincipal();
+		String username = userDetail.getUsername();
+		Vendedor vendedor = this.vendedorService.findVendedorByUsername(username);
+		
+		if (result.hasErrors()) {
 			return "/peliculas/formCreatePeliculas";
-		}else {
+		} else {
 			this.peliculaService.savePelicula(pel);
-			return "redirect:/peliculas/" + pel.getId();
+			if (vendedor.getPeliculas() == null) {
+				Collection<Pelicula> peliculas = new ArrayList<Pelicula>();
+				peliculas.add(pel);
+				vendedor.getPeliculas().addAll(peliculas);
+			} else {
+				vendedor.getPeliculas().add(pel);
+			}
+			this.vendedorService.save(vendedor);
+			return mostrarProductos(model);
 		}
 	}
+
 	
-	//borra una determinada pelicula
 	@GetMapping("/peliculas/delete/{peliculaId}")
 	public String deletePelicula(@PathVariable("peliculaId") int peliculaId, final ModelMap modelMap) {
+
 		Pelicula peliculaBorrar = this.peliculaService.findPeliculaById(peliculaId);
-		String view = "peliculas/PeliculasList";
-		if (peliculaBorrar != null) {
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		UserDetails userDetail = (UserDetails) auth.getPrincipal();
+		String username = userDetail.getUsername();
+		Vendedor vendedor = this.vendedorService.findVendedorByUsername(username);
+		Collection<Pelicula> peliculas = vendedor.getPeliculas();
+
+		if ((peliculaBorrar != null) && (peliculas.contains(peliculaBorrar))) {
+			peliculas.remove(peliculaBorrar);
+			this.vendedorService.save(vendedor);
 			this.peliculaService.delete(peliculaBorrar);
-			view = this.showPeliculasList(modelMap);
-		}else {
+			modelMap.addAttribute("message", "¡Producto eliminado!");
+			return mostrarProductos(modelMap);
+		} else {
 			modelMap.addAttribute("message", "ERROR!");
+			return "peliculas/PeliculasList";
 		}
-		
-		return view;
-		
 	}
 	
+
+	
+	public String mostrarProductos(ModelMap modelMap) {
+		
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		UserDetails userDetail = (UserDetails) auth.getPrincipal();
+		String usuario = userDetail.getUsername();
+		Vendedor vendedor = this.vendedorService.findVendedorByUsername(usuario);
+		
+		
+		if(this.vendedorService.obtenerPeliculas(vendedor.getId())!=null) {
+			List<Pelicula> peliculas = new ArrayList<>();
+			peliculas.addAll(this.vendedorService.obtenerPeliculas(vendedor.getId()));
+			modelMap.addAttribute("peliculas", peliculas);
+		}
+		if(this.vendedorService.obtenerVideojuegos(vendedor.getId())!=null) {
+			List<Videojuego> videojuegos = new ArrayList<>();
+			videojuegos.addAll(this.vendedorService.obtenerVideojuegos(vendedor.getId()));
+			modelMap.addAttribute("videojuegos", videojuegos);
+		}
+		if(this.vendedorService.obtenerMerchandasings(vendedor.getId())!=null) {
+			List<Merchandasing> merch = new ArrayList<>();
+			merch.addAll(this.vendedorService.obtenerMerchandasings(vendedor.getId()));
+			modelMap.addAttribute("merch", merch);
+		}
+		
+		return "/productos/productosVendedor";
+	}
+
 }
